do not try to process applets list if no one received

+++ NetRocks/src/Protocol/SSH/ProtocolSCP.cpp
@@ -634,6 +634,7 @@ ProtocolSCP::ProtocolSCP(const std::string &host, unsigned int port,
 	}
 
 	bool busybox = false;
+	bool applets_list = false;
 	if (cmd.Execute("readlink /bin/sh") == 0) {
 		if (cmd.Output().find("busybox") != std::string::npos && cmd.Execute("busybox") == 0) {
 			fprintf(stderr, "ProtocolSCP: BusyBox detected\n");
@@ -649,21 +650,24 @@ ProtocolSCP::ProtocolSCP(const std::string &host, unsigned int port,
 			// Its enough arguments to assume that ls will be handled by busybox.
 			fprintf(stderr, "ProtocolSCP: BusyBox assumed\n");
 			busybox = true;
+			applets_list = (busybox_return_code == 0);
 		}
 	}
 
 	if (busybox) {
-		// some busybox systems may miss very usual things, analyze busybox command output
-		// where it printed lit of supported commands
-		std::vector<std::string> words;
-		StrExplode(words, cmd.Output(), "\t ,");
-		if (std::find(words.begin(), words.end(), _quirks.rm_file) == words.end()) {
-			fprintf(stderr, "ProtocolSCP: '%s' unsupported\n", _quirks.rm_file);
-			_quirks.rm_file = "rm -f";
-		}
-		if (std::find(words.begin(), words.end(), _quirks.rm_dir) == words.end()) {
-			fprintf(stderr, "ProtocolSCP: '%s' unsupported\n", _quirks.rm_dir);
-			_quirks.rm_dir = "rm -f -d";
+		if (applets_list) {
+			// some busybox systems may miss very usual things, analyze busybox command output
+			// where it printed lit of supported commands
+			std::vector<std::string> words;
+			StrExplode(words, cmd.Output(), "\t ,");
+			if (std::find(words.begin(), words.end(), _quirks.rm_file) == words.end()) {
+				fprintf(stderr, "ProtocolSCP: '%s' unsupported\n", _quirks.rm_file);
+				_quirks.rm_file = "rm -f";
+			}
+			if (std::find(words.begin(), words.end(), _quirks.rm_dir) == words.end()) {
+				fprintf(stderr, "ProtocolSCP: '%s' unsupported\n", _quirks.rm_dir);
+				_quirks.rm_dir = "rm -f -d";
+			}
 		}
 		_quirks.ls_supports_dash_f = false;
 	}
